Backend/requirements.txt
# langchain==0.0.154
langchain
openai
PyPDF2
python-dotenv
faiss-cpu==1.7.4
tiktoken 
flask
qdrant-client